refactor: rename polyhedron/traits.rs to geometry.rs and fix doc tests
- Rename traits.rs to geometry.rs (no traits, contains geometric types)
- Update all import references to use new module name
- Fix README doc test compilation errors with proper imports and unwrap()
- Convert ignored doc tests to runnable examples with concrete values
- All 44 tests now passing (10 doc tests previously ignored)
