+++ README.md
@@ -80,16 +80,16 @@ Make sure DGGRID is compiled and available on your system. Remember the path whe
 ## Usage Example
 
 Create a new crate with `cargo new` and add this dependency in your `cargo.toml`. I expect to publish this to crates.io in the future, which will simplify this with `cargo add dggrs`.
-````rust
+```toml
 [dependencies]
 dggrs = {version = "0.1.0", git = git@gitlab.com/geoinsight/dggrs.git}
 ```
 
 In your `main.rs` add the following code. In this example the DGGRID generator service is instantiated using the path to the DGGRID executable `dggrid` and a path to the work directory `/dev/shm`. 
 
-````rust
+```rust
 use geo::geometry::Point;
-use geoplegma::dggrs;
+use dggrs;
 fn main() {
     let configs = vec![
         // (
@@ -118,10 +118,10 @@ fn main() {
     for (tool, dggs, zone_id) in configs {
         println!("=== DGGS Type: {} ===", dggs);
 
-        let generator = dggrs::get(&tool, &dggs);
+        let generator = dggrs::get(&tool, &dggs).unwrap();
 
         println!("Global");
-        let result = generator.zones_from_bbox(2, false, None);
+        let result = generator.zones_from_bbox(2, false, None).unwrap();
         println!(
             "{:?} \nGenerated {} zones",
             result.zones,
@@ -129,7 +129,7 @@ fn main() {
         );
 
         println!("Global with Bbox");
-        let result = generator.zones_from_bbox(2, false, bbox.clone());
+        let result = generator.zones_from_bbox(2, false, bbox.clone()).unwrap();
         println!(
             "{:?} \nGenerated {} zones",
             result.zones,
@@ -137,7 +137,7 @@ fn main() {
         );
 
         println!("Point");
-        let result = generator.zone_from_point(6, pnt, false);
+        let result = generator.zone_from_point(6, pnt, false).unwrap();
         println!(
             "{:?} \nGenerated {} zones",
             result.zones,
@@ -145,7 +145,7 @@ fn main() {
         );
 
         println!("Subzones of {}", zone_id);
-        let result = generator.zones_from_parent(6, zone_id.clone(), false);
+        let result = generator.zones_from_parent(6, zone_id.clone(), false).unwrap();
         println!(
             "{:?} \nGenerated {} zones",
             result.zones,
@@ -153,7 +153,7 @@ fn main() {
         );
 
         println!("Single Zone {}", zone_id.clone());
-        let result = generator.zone_from_id(zone_id.clone(), false);
+        let result = generator.zone_from_id(zone_id.clone(), false).unwrap();
         println!(
             "{:?} \nGenerated {} zones",
             result.zones,
@@ -162,7 +162,7 @@ fn main() {
     }
 }
 
-````
+```
 
 Instead of printing out the length of `result.zones.len()` you can also print out the struct itself.
