+++ gp-proj/src/projections/polyhedron/geometry.rs
