+++ gp-proj/src/projections/polyhedron/mod.rs
@@ -8,10 +8,10 @@
 // except according to those terms
 
 pub mod icosahedron;
-pub mod traits;
+pub mod geometry;
 pub mod polyhedron;
 pub mod spherical_geometry;
 
 // Re-export the main types
 pub use polyhedron::Polyhedron;
-pub use traits::{Face, ArcLengths};
+pub use geometry::{Face, ArcLengths};
